:construction: migrated the api protocol completely and re-drafted the source plugins spec in the right folder
