+++ server/controllers/ConnectionController.js
@@ -1,80 +1,8 @@
-const querystring = require("querystring");
-const moment = require("moment");
 const fs = require("fs");
 
 const db = require("../models/models");
 const ProjectController = require("./ProjectController");
-const paginateRequests = require("../modules/paginateRequests");
-const safeRequest = require("../modules/safeRequest");
-const determineType = require("../modules/determineType");
-const drCacheController = require("./DataRequestCacheController");
-const { applyApiVariables } = require("../modules/applyVariables");
-const { buildChartRuntimeContext } = require("../modules/chartRuntimeFilters");
-const {
-  getItemCount,
-  sanitizeSnippet,
-  serializeResponsePreview,
-} = require("../modules/updateAudit");
-const {
-  checkAndGetCache,
-  completeConnectorAudit,
-  failConnectorAudit,
-} = require("../sources/shared/connectorRuntime");
-
-const getMomentObj = (timezone) => {
-  if (timezone) {
-    return (...args) => moment(...args).tz(timezone);
-  } else {
-    return (...args) => moment.utc(...args);
-  }
-};
-
-function buildApiPolicyContext(source, connection, overrides = {}) {
-  const context = {
-    source,
-    teamId: overrides.teamId || connection?.team_id || null,
-    connectionId: overrides.connectionId || connection?.id || null,
-  };
-
-  if (typeof overrides.allowPrivateHost === "boolean") {
-    context.allowPrivateHost = overrides.allowPrivateHost;
-  } else if (overrides.allowPrivateHost === null) {
-    context.allowPrivateHost = null;
-  } else if (typeof connection?.allowPrivateHost === "boolean") {
-    context.allowPrivateHost = connection.allowPrivateHost;
-  } else {
-    context.allowPrivateHost = null;
-  }
-
-  return context;
-}
-
-function isArrayPresent(responseData) {
-  let arrayFound = false;
-  Object.keys(responseData).forEach((k1) => {
-    if (determineType(responseData[k1]) === "array") {
-      arrayFound = true;
-    }
-
-    if (!arrayFound && determineType(responseData[k1]) === "object") {
-      Object.keys(responseData[k1]).forEach((k2) => {
-        if (determineType(responseData[k1][k2]) === "array") {
-          arrayFound = true;
-        }
-
-        if (!arrayFound && determineType(responseData[k1][k2]) === "object") {
-          Object.keys(responseData[k1][k2]).forEach((k3) => {
-            if (determineType(responseData[k1][k2][k3]) === "array") {
-              arrayFound = true;
-            }
-          });
-        }
-      });
-    }
-  });
-
-  return arrayFound;
-}
+const apiProtocol = require("../sources/shared/protocols/api.protocol");
 
 class ConnectionController {
   constructor() {
@@ -240,110 +168,35 @@ class ConnectionController {
   }
 
   getApiTestOptions(connection) {
-    if (connection.type !== "api") return false;
-
-    const testOptions = {
-      url: connection.host,
-      method: "GET",
-      headers: {},
-      resolveWithFullResponse: true,
-    };
-
-    let globalHeaders = connection.options;
-    if (connection.getHeaders) {
-      globalHeaders = connection.getHeaders(connection);
-    } else if (connection.authentication && connection.authentication.type === "bearer_token") {
-      testOptions.headers.authorization = `Bearer ${connection.authentication.token}`;
-    }
-
-    if (globalHeaders && globalHeaders.length > 0) {
-      for (const option of connection.options) {
-        testOptions.headers[Object.keys(option)[0]] = option[Object.keys(option)[0]];
-      }
-    }
-
-    // Basic Auth
-    if (connection.authentication && connection.authentication.type === "basic_auth") {
-      testOptions.auth = {
-        user: connection.authentication.user,
-        pass: connection.authentication.pass,
-      };
-    }
-
-    return testOptions;
+    return apiProtocol.getApiTestOptions(connection);
   }
 
   testRequest(data, extras) {
-    const certificates = {};
-    if (extras?.files?.length > 0) {
-      try {
-        extras.files.forEach((file) => {
-          // Handle SSL certificates
-          if (file.fieldname === "sslCa" || file.fieldname === "sslCert" || file.fieldname === "sslKey") {
-            certificates[file.fieldname] = file.path; // Use the temporary file path for testing
-          }
-          // Handle SSH private key
-          if (file.fieldname === "sshPrivateKey") {
-            certificates.sshPrivateKey = file.path;
-          }
-        });
-      } catch (error) {
-        return Promise.reject(new Error(`Error processing certificate files: ${error.message}`));
-      }
-    }
-
-    let connectionParams = { ...data };
-
-    if (Object.keys(certificates).length > 0) {
-      connectionParams = { ...connectionParams, ...certificates };
-    }
-
-    if (data.type === "api") {
-      return this.testApi(connectionParams, buildApiPolicyContext(
-        "connection_type_test",
-        null,
-        {
-          teamId: data.team_id || null,
-          connectionId: null,
-          // Unsaved test payloads should not override private network policy.
-          allowPrivateHost: null,
+    try {
+      const certificates = {};
+      extras?.files?.forEach((file) => {
+        if (["sslCa", "sslCert", "sslKey", "sshPrivateKey"].includes(file.fieldname)) {
+          certificates[file.fieldname] = file.path;
         }
-      ));
-    }
+      });
 
-    return new Promise((resolve, reject) => reject(new Error("No request type specified")));
+      return apiProtocol.testUnsavedConnection({
+        connection: { ...data, ...certificates },
+        extras,
+      });
+    } catch (error) {
+      return Promise.reject(new Error(`Error processing certificate files: ${error.message}`));
+    }
   }
 
   testApi(data, policyContext = {}) {
-    const testOpt = this.getApiTestOptions(data);
-    return safeRequest(testOpt, policyContext);
+    return apiProtocol.testApi(data, policyContext);
   }
 
   testConnection(id) {
-    let gConnection;
     return db.Connection.findByPk(id)
       .then((connection) => {
-        gConnection = connection;
-        switch (connection.type) {
-          case "api":
-            return this.testApi(connection, buildApiPolicyContext("connection_test", connection));
-          default:
-            return new Promise((resolve, reject) => reject(new Error(400)));
-        }
-      })
-      .then((response) => {
-        switch (gConnection.type) {
-          case "api":
-            if (response.statusCode < 300) {
-              return new Promise((resolve) => resolve({ success: true }));
-            }
-            return new Promise((resolve, reject) => reject(new Error(400)));
-          default:
-            return new Promise((resolve, reject) => reject(new Error(400)));
-        }
-      })
-      .then(() => {
-        return new Promise((resolve) => resolve({ success: true }));
+        return apiProtocol.testConnection({ connection });
       })
       .catch((err) => {
         return new Promise((resolve, reject) => reject(err));
@@ -353,79 +206,17 @@ class ConnectionController {
   testApiRequest({
     connection_id, dataRequest, itemsLimit, items, offset, pagination, paginationField,
   }) {
-    const limit = itemsLimit
-      ? parseInt(itemsLimit, 10) : 0;
     return this.findById(connection_id)
       .then((connection) => {
-        const policyContext = buildApiPolicyContext("api_request_test", connection);
-        const tempUrl = `${connection.getApiUrl(connection)}${dataRequest.route || ""}`;
-        const queryParams = querystring.parse(tempUrl.split("?")[1]);
-
-        let url = tempUrl;
-        if (url.indexOf("?") > -1) {
-          url = tempUrl.substring(0, tempUrl.indexOf("?"));
-        }
-
-        const options = {
-          url,
-          method: dataRequest.method || "GET",
-          headers: {},
-          qs: queryParams,
-          resolveWithFullResponse: true,
-          simple: false,
-        };
-
-        // prepare the headers
-        let headers = {};
-        if (dataRequest.useGlobalHeaders) {
-          const globalHeaders = connection.getHeaders(connection);
-          for (const opt of globalHeaders) {
-            headers = Object.assign(opt, headers);
-          }
-
-          if (dataRequest.headers) {
-            headers = Object.assign(dataRequest.headers, headers);
-          }
-        }
-
-        options.headers = headers;
-
-        if (dataRequest.body && dataRequest.method !== "GET") {
-          options.body = dataRequest.body;
-          options.headers["Content-Type"] = "application/json";
-        }
-
-        if (pagination) {
-          if ((options.url.indexOf(`?${items}=`) || options.url.indexOf(`&${items}=`))
-            && (options.url.indexOf(`?${offset}=`) || options.url.indexOf(`&${offset}=`))
-          ) {
-            return paginateRequests(dataRequest.template, {
-              options,
-              limit,
-              items,
-              offset,
-              paginationField,
-              policyContext,
-            });
-          }
-        }
-
-        return safeRequest(options, policyContext);
-      })
-      .then((response) => {
-        if (pagination) {
-          return new Promise((resolve) => resolve(response));
-        }
-
-        if (response.statusCode < 300) {
-          try {
-            return new Promise((resolve) => resolve(JSON.parse(response.body)));
-          } catch (e) {
-            return new Promise((resolve, reject) => reject(400));
-          }
-        } else {
-          return new Promise((resolve, reject) => reject(response.statusCode));
-        }
+        return apiProtocol.previewDataRequest({
+          connection,
+          dataRequest,
+          itemsLimit,
+          items,
+          offset,
+          pagination,
+          paginationField,
+        });
       })
       .catch((error) => {
         return new Promise((resolve, reject) => reject(error));
@@ -433,412 +224,25 @@ class ConnectionController {
   }
 
   async runApiRequest(id, chartId, dataRequest, getCache, filters, timezone = "", runtimeVariables = {}, auditContext = null) {
-    if (getCache) {
-      const drCache = await checkAndGetCache(id, dataRequest);
-      if (drCache) {
-        await completeConnectorAudit(auditContext, {
-          cacheHit: true,
-          connectionType: "api",
-          ...serializeResponsePreview(drCache.responseData),
-        });
-        return drCache;
-      }
-    }
-
-    const limit = dataRequest.itemsLimit
-      ? parseInt(dataRequest.itemsLimit, 10) : 0;
-    const { variables } = dataRequest;
-
-    return this.findById(id)
-      .then(async (connection) => {
-        const policyContext = buildApiPolicyContext("api_request_run", connection);
-        // Apply variable substitution for API requests
-        let processedRoute = dataRequest.route || "";
-        let processedHeaders = dataRequest.headers || {};
-        let processedBody = dataRequest.body || "";
-
-        try {
-          const result = applyApiVariables(dataRequest, runtimeVariables);
-          processedRoute = result.processedRoute || processedRoute;
-          processedHeaders = result.processedHeaders || processedHeaders;
-          processedBody = result.processedBody || processedBody;
-        } catch (error) {
-          // If there's an error in variable processing, return it
-          return Promise.reject(error);
-        }
-
-        let tempUrl = connection.getApiUrl(connection);
-        let route = processedRoute;
-        if (route && (route[0] !== "/" && route[0] !== "?")) {
-          route = `/${route}`;
-        }
-
-        tempUrl += route;
-
-        const queryParams = querystring.parse(tempUrl.split("?")[1]);
-
-        // if any queryParams has variables, modify them here
-        if (queryParams && Object.keys(queryParams).length > 0) {
-          // First, process generic variables (excluding start_date and end_date)
-          if (dataRequest.VariableBindings && dataRequest.VariableBindings.length > 0) {
-            // Process each query parameter for variables
-            for (const q of Object.keys(queryParams)) {
-              const paramValue = queryParams[q];
-              if (typeof paramValue === "string") {
-                let processedValue = paramValue;
-
-                // Find all variables in this parameter value
-                const variableMatches = [...paramValue.matchAll(/\{\{([^}]+)\}\}/g)];
-
-                for (const match of variableMatches) {
-                  const variableName = match[1].trim();
-
-                  // Skip reserved date variables - they're handled separately below
-                  if (variableName === "start_date" || variableName === "end_date") {
-                    // oxlint-disable-next-line no-continue
-                    continue;
-                  }
-
-                  const binding = dataRequest.VariableBindings
-                    .find((vb) => vb.name === variableName);
-
-                  // Check for runtime variable value first
-                  const runtimeValue = runtimeVariables[variableName];
-                  const hasRuntimeValue = runtimeValue !== null
-                    && runtimeValue !== undefined && runtimeValue !== "";
-
-                  // Check for default value
-                  const hasDefaultValue = binding?.default_value !== null
-                    && binding?.default_value !== undefined
-                    && binding?.default_value !== "";
-
-                  if (hasRuntimeValue) {
-                    // Priority 1: Use runtime value
-                    let replacementValue = runtimeValue;
-
-                    // Handle different data types based on binding type (if available)
-                    if (binding?.type) {
-                      switch (binding.type) {
-                        case "string":
-                          replacementValue = String(runtimeValue);
-                          break;
-                        case "number":
-                          replacementValue = Number.isNaN(Number(runtimeValue))
-                            ? "0" : String(runtimeValue);
-                          break;
-                        case "boolean":
-                          replacementValue = (runtimeValue === "true" || runtimeValue === true)
-                            ? "true" : "false";
-                          break;
-                        case "date":
-                          replacementValue = String(runtimeValue);
-                          break;
-                        default:
-                          replacementValue = String(runtimeValue);
-                      }
-                    } else {
-                      // No binding type info, treat as string
-                      replacementValue = String(runtimeValue);
-                    }
-
-                    processedValue = processedValue.replace(match[0], replacementValue);
-                  } else if (hasDefaultValue && binding) {
-                    // Priority 2: Use default value
-                    let replacementValue = binding.default_value;
-
-                    if (binding.type) {
-                      switch (binding.type) {
-                        case "string":
-                          replacementValue = String(binding.default_value);
-                          break;
-                        case "number":
-                          replacementValue = Number.isNaN(Number(binding.default_value))
-                            ? "0" : String(binding.default_value);
-                          break;
-                        case "boolean":
-                          replacementValue = binding.default_value === "true"
-                            || binding.default_value === true ? "true" : "false";
-                          break;
-                        case "date":
-                          replacementValue = String(binding.default_value);
-                          break;
-                        default:
-                          replacementValue = String(binding.default_value);
-                      }
-                    } else {
-                      replacementValue = String(binding.default_value);
-                    }
-
-                    processedValue = processedValue.replace(match[0], replacementValue);
-                  } else {
-                    // Priority 3: No runtime value and no default value
-                    if (binding?.required) {
-                      // Required variable without value - throw error
-                      const errorMsg = `Required variable '${variableName}' has no value provided and no default value`;
-                      throw new Error(errorMsg);
-                    }
-
-                    // Not required and no value - remove the placeholder
-                    processedValue = processedValue.replace(match[0], "");
-                  }
-                }
-
-                // Update the query parameter with processed value
-                queryParams[q] = processedValue;
-              }
-            }
-          }
-
-          // Now handle special date variables
-          // first, check for the keys to avoid making an unnecessary query to the DB
-          const keysFound = {};
-          Object.keys(queryParams).forEach((q) => {
-            const paramValue = queryParams[q];
-            // Check for exact matches
-            if (paramValue === "{{start_date}}") {
-              keysFound[q] = { type: "startDate", format: "single" };
-            } else if (paramValue === "{{end_date}}") {
-              keysFound[q] = { type: "endDate", format: "single" };
-            } else if (typeof paramValue === "string") {
-              // Check for combined variables using regex
-              const startDateMatch = paramValue.match(/{{start_date}}/);
-              const endDateMatch = paramValue.match(/{{end_date}}/);
-              if (startDateMatch || endDateMatch) {
-                keysFound[q] = {
-                  type: "combined",
-                  hasStartDate: !!startDateMatch,
-                  hasEndDate: !!endDateMatch,
-                  originalValue: paramValue
-                };
-              }
-            }
-          });
-
-          // something was found, go through all and replace the date variables
-          if (Object.keys(keysFound).length > 0) {
-            const chart = await db.Chart.findByPk(chartId);
-            const runtimeContext = chart
-              ? buildChartRuntimeContext(chart, filters, runtimeVariables, timezone)
-              : null;
-            const effectiveDateRange = runtimeContext?.effectiveDateRange;
-
-            if (chart && effectiveDateRange) {
-              Object.keys(keysFound).forEach((q) => {
-                const value = keysFound[q];
-                const startDate = effectiveDateRange.startDate;
-                const endDate = effectiveDateRange.endDate;
-
-                if (value.format === "single") {
-                  if (value.type === "startDate" && startDate) {
-                    queryParams[q] = startDate.format(chart.dateVarsFormat || "");
-                  } else if (value.type === "endDate" && endDate) {
-                    queryParams[q] = endDate.format(chart.dateVarsFormat || "");
-                  }
-                } else if (value.type === "combined") {
-                  let formattedValue = value.originalValue;
-                  if (value.hasStartDate && startDate) {
-                    formattedValue = formattedValue.replace(/{{start_date}}/g, startDate.format(chart.dateVarsFormat || ""));
-                  }
-                  if (value.hasEndDate && endDate) {
-                    formattedValue = formattedValue.replace(/{{end_date}}/g, endDate.format(chart.dateVarsFormat || ""));
-                  }
-                  queryParams[q] = formattedValue;
-                }
-              });
-            } else if (variables?.startDate?.value && variables?.endDate?.value) {
-              Object.keys(keysFound).forEach((q) => {
-                const value = keysFound[q];
-                const startDate = getMomentObj(timezone)(variables.startDate.value);
-                const endDate = getMomentObj(timezone)(variables.endDate.value);
-
-                if (value.format === "single") {
-                  if (value.type === "startDate" && startDate) {
-                    queryParams[q] = startDate.format(variables.dateFormat?.value || "");
-                  } else if (value.type === "endDate" && endDate) {
-                    queryParams[q] = endDate.format(variables.dateFormat?.value || "");
-                  }
-                } else if (value.type === "combined") {
-                  let formattedValue = value.originalValue;
-                  if (value.hasStartDate && startDate) {
-                    formattedValue = formattedValue.replace(/{{start_date}}/g, startDate.format(variables.dateFormat?.value || ""));
-                  }
-                  if (value.hasEndDate && endDate) {
-                    formattedValue = formattedValue.replace(/{{end_date}}/g, endDate.format(variables.dateFormat?.value || ""));
-                  }
-                  queryParams[q] = formattedValue;
-                }
-              });
-            }
-          }
-        }
-
-        let url = tempUrl;
-        if (url.indexOf("?") > -1) {
-          url = tempUrl.substring(0, tempUrl.indexOf("?"));
-        }
-
-        // if ant variable queryParams are left, remove them
-        if (queryParams && Object.keys(queryParams).length > 0) {
-          Object.keys(queryParams).forEach((q) => {
-            if (queryParams[q] === "{{start_date}}" || queryParams[q] === "{{end_date}}") {
-              delete queryParams[q];
-            }
-          });
-        }
-
-        const options = {
-          url,
-          method: dataRequest.method || "GET",
-          headers: {},
-          qs: queryParams,
-          resolveWithFullResponse: true,
-          simple: false,
-        };
-
-        // prepare the headers
-        let headers = {};
-        if (dataRequest.useGlobalHeaders) {
-          const globalHeaders = connection.getHeaders(connection);
-          for (const opt of globalHeaders) {
-            headers = Object.assign(opt, headers);
-          }
-
-          if (processedHeaders) {
-            headers = Object.assign(processedHeaders, headers);
-          }
-        }
-
-        options.headers = headers;
-
-        if (processedBody && dataRequest.method !== "GET") {
-          options.body = processedBody;
-          options.headers["Content-Type"] = "application/json";
-        }
-
-        // Basic auth
-        if (connection.authentication && connection.authentication.type === "basic_auth") {
-          options.auth = {
-            user: connection.authentication.user,
-            pass: connection.authentication.pass,
-          };
-        }
-
-        if (dataRequest.pagination) {
-          if ((options.url.indexOf(`?${dataRequest.items}=`) || options.url.indexOf(`&${dataRequest.items}=`))
-            && (options.url.indexOf(`?${dataRequest.offset}=`) || options.url.indexOf(`&${dataRequest.offset}=`))
-          ) {
-            return paginateRequests(dataRequest.template, {
-              options,
-              limit,
-              items: dataRequest.items,
-              offset: dataRequest.offset,
-              paginationField: dataRequest.paginationField,
-              policyContext,
-            });
-          }
-        }
-
-        return safeRequest(options, policyContext);
-      })
-      .then(async (response) => {
-        if (dataRequest.pagination) {
-          // cache the data for later use
-          const dataToCache = {
-            dataRequest,
-            responseData: {
-              data: response,
-            },
-            connection_id: id,
-          };
-
-          await drCacheController.create(dataRequest.id, dataToCache);
-          await completeConnectorAudit(auditContext, {
-            cacheHit: false,
-            connectionType: "api",
-            paginated: true,
-            itemCount: getItemCount(response),
-            responseSnippet: sanitizeSnippet(response),
-          });
-
-          return new Promise((resolve) => resolve(dataToCache));
-        }
-
-        if (response.statusCode < 300) {
-          try {
-            let responseData = JSON.parse(response.body);
-
-            // check if there are arrays to take into account
-            // transform the data in 1-item array if that's the case
-            // check for arrays in 3 levels
-            if (determineType(responseData) === "object" && !isArrayPresent(responseData)) {
-              responseData = [responseData];
-            }
-
-            // cache the data for later use
-            const dataToCache = {
-              dataRequest,
-              responseData: {
-                data: responseData,
-              },
-              connection_id: id,
-            };
-
-            await drCacheController.create(dataRequest.id, dataToCache);
-            await completeConnectorAudit(auditContext, {
-              cacheHit: false,
-              connectionType: "api",
-              statusCode: response.statusCode,
-              bodySnippet: sanitizeSnippet(response.body),
-              ...serializeResponsePreview(dataToCache.responseData),
-            });
-
-            return new Promise((resolve) => resolve(dataToCache));
-          } catch (e) {
-            return new Promise((resolve, reject) => reject(400));
-          }
-        } else {
-          return new Promise((resolve, reject) => reject(response.statusCode));
-        }
-      })
-      .catch(async (error) => {
-        await failConnectorAudit(auditContext, error, error.auditStage || "connection", {
-          cacheHit: false,
-          connectionType: "api",
-          statusCode: error?.statusCode || (typeof error === "number" ? error : null),
-          responseSnippet: sanitizeSnippet(error?.body || error?.error || error?.message || error),
-        });
-        return new Promise((resolve, reject) => reject(error));
-      });
+    const connection = await this.findById(id);
+    return apiProtocol.runDataRequest({
+      connection,
+      dataRequest,
+      chartId,
+      getCache,
+      filters,
+      timezone,
+      variables: runtimeVariables,
+      auditContext,
+    });
   }
 
   async getApiBuilderMetadata(connectionId, { includeSensitive = false } = {}) {
     const connection = await this.findById(connectionId);
-    let globalHeaders = connection.getHeaders(connection);
-
-    if (!Array.isArray(globalHeaders)) {
-      try {
-        globalHeaders = JSON.parse(globalHeaders);
-      } catch (error) {
-        globalHeaders = [];
-      }
-    }
-
-    return {
-      host: connection.getApiUrl(connection),
-      globalHeaders: globalHeaders.map((header) => {
-        const key = Object.keys(header || {})[0];
-        const value = key ? header[key] : "";
-        let serializedValue = value;
-        if (!includeSensitive) {
-          serializedValue = value ? "Hidden" : "";
-        }
-        return {
-          key: key || "",
-          value: serializedValue,
-        };
-      }),
-      hasGlobalHeaders: globalHeaders.length > 0,
-    };
+    return apiProtocol.getBuilderMetadata({
+      connection,
+      options: { includeSensitive },
+    });
   }
 
   async duplicateConnection(connectionId, name) {
